Barebones of Project (4/19)
I started with formatting the menu and how I think the menu of the shop will look.

I added empty functions to so that we can start filling them out. Once I can see how the different functions will work I can start placing them in the right spots.

I didn't add anything in the shop function because I think as a customer they should only be able to view the shop but if there is a specific account we want to add to be able to edit the items so we can use the items class.

Just an idea and writing it all down so I don't forget!