Update: Inventory and Shop completed (4/24)
I did sqlite stuff for these classes so you dont need to worry about these ones!

Def need some thought on how to do the cart but the ITEM function creates a shop db so it should have all the items in it that are added from the inventory!

Ill think on how to do the log in just let me know what we decide to do!